Move GridLengthAnimation to its own folder
